# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.FailureMsg:
    properties:
      error:
        type: string
    type: object
...
  model.Role:
    properties:
      Id:
        type: integer
      creationDate:
        type: string
      description:
        type: string
      roleName:
        type: string
    type: object
  model.RolesList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Role'
        type: array
    type: object
  model.StorageVolume:
    properties:
      Id:
        type: integer
      creationDate:
        type: string
      creatorId:
        type: integer
      deviceId:
        type: string
      deviceModel:
        type: string
      mountPoint:
        type: string
      storageType:
        type: string
      systemId:
        type: integer
      volumeFormat:
        type: string
      volumeLabel:
        type: string
      volumeName:
        type: string
      volumeSize:
        type: integer
    type: object
  model.StorageVolumes:
    properties:
      volumes:
        items:
          $ref: '#/definitions/model.StorageVolume'
        type: array
    type: object
  model.SuccessMsg:
    properties:
      message:
        type: string
    type: object
...
host: localhost:5000
info:
  contact:
    name: Gary Greene
    url: https://github.com/greeneg/allocatord
  description: An API for managing OS imaging
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Allocator Daemon
  version: 0.1.5
paths:
...
  /role:
    post:
      consumes:
      - application/json
      description: Add a new role
      parameters:
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register role
      tags:
      - role
  /role/{roleId}:
    delete:
      consumes:
      - application/json
      description: Delete a role
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete role
      tags:
      - role
  /role/byId/{roleId}:
    get:
      description: Retrieve a role by its Id
      parameters:
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a role by its Id
      tags:
      - role
  /role/byName/{roleName}:
    get:
      description: Retrieve a role by its role name
      parameters:
      - description: Role Name
        in: path
        name: roleName
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a role by its role name
      tags:
      - role
  /roles:
    get:
      description: Retrieve list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RolesList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve list of all roles
      tags:
      - role
  /storageVolume:
    post:
      consumes:
      - application/json
      description: Add a new storage volume
      parameters:
      - description: Storage Volume data
        in: body
        name: storageVolume
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register storage volume
      tags:
      - storage-volumes
  /storageVolume/{storageVolumeId}:
    delete:
      consumes:
      - application/json
      description: Delete a storage volume
      parameters:
      - description: Storage Volume Id
        in: path
        name: storageVolumeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete storage volume
      tags:
      - storage-volumes
  /storageVolume/{systemId}/byLabel/{storageVolumeLabel}:
    get:
      description: Retrieve a storage volume by its volume label
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Storage Volume Label
        in: path
        name: storageVolumeLabel
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageVolume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a storage volume by its volume label
      tags:
      - storage-volumes
  /storageVolume/byId/{storageVolumeId}:
    get:
      description: Retrieve a storage volume by its Id
      parameters:
      - description: Storage Volume ID
        in: path
        name: storageVolumeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageVolume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a storage volume by its Id
      tags:
      - storage-volumes
  /storageVolumes:
    get:
      description: Retrieve list of all storage volumes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageVolumes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve list of all storage volumes
      tags:
      - storage-volumes
  /storageVolumes/{systemId}:
    get:
      description: Retrieve storage volumes by system Id
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageVolumes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve storage volumes by system Id
      tags:
      - storage-volumes
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"